feat(6697): extract sdk direct inline tests
