+++ crates/kamn-e2e-harness/src/drivers/sdk_direct.rs
@@ -1641,908 +1641,5 @@ fn validate_s14_proof_response(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        live_execution_enabled_from_env, parse_s15_budget_env_u128, run_live_s01_discovery_probe,
-        run_live_s02_direct_message_probe, run_live_s03_group_channel_probe,
-        run_live_s04_task_lifecycle_probe, run_live_s05_escrow_settlement_probe,
-        run_live_s06_proof_verification_probe, run_live_s07_replay_protection_probe,
-        run_live_s08_crash_recovery_probe, run_live_s09_transport_failover_probe,
-        run_live_s10_topology_coherence_probe, run_live_s11_signer_rotation_probe,
-        run_live_s12_retention_deletion_probe, run_live_s13_bridge_forwarding_probe,
-        run_live_s14_batch_merkle_probe, run_live_s15_performance_smoke_probe,
-        validate_live_s03_list_messages_response, validate_live_s03_query_message_response,
-        validate_live_s05_release_escrow_receipt, validate_s07_replay_reason_marker,
-        validate_s08_distinct_message_ids, validate_s08_message_receipt_fields,
-        validate_s08_query_message_response, validate_s12_content_field_coherence,
-        validate_s12_content_id_match, validate_s13_bridge_field_coherence,
-        validate_s13_bridge_id_match, validate_s14_proof_response,
-        validate_s15_latency_budget_samples, SdkDirectDriver, SDK_DIRECT_LIVE_ENV,
-    };
-    use std::env;
-    use std::ffi::OsString;
-    use std::sync::PoisonError;
-
-    fn with_env_vars<F>(updates: &[(&str, Option<&str>)], test: F)
-    where
-        F: FnOnce(),
-    {
-        let _guard = crate::drivers::test_env_lock()
-            .lock()
-            .unwrap_or_else(PoisonError::into_inner);
-        let previous = updates
-            .iter()
-            .map(|(key, _)| ((*key).to_owned(), env::var_os(key)))
-            .collect::<Vec<(String, Option<OsString>)>>();
-
-        for (key, value) in updates {
-            match value {
-                Some(value) => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::set_var(key, value) }
-                }
-                None => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::remove_var(key) }
-                }
-            }
-        }
-
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(test));
-        for (key, value) in previous {
-            match value {
-                Some(value) => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::set_var(key, value) }
-                }
-                None => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::remove_var(key) }
-                }
-            }
-        }
-        if let Err(payload) = result {
-            std::panic::resume_unwind(payload);
-        }
-    }
-
-    #[test]
-    fn unit_parse_bool_flag_accepts_true_like_values() {
-        for value in ["1", "true", "TRUE", "yes", "on"] {
-            assert!(
-                crate::drivers::shared_helpers::parse_bool_flag(value),
-                "expected truthy for {value}"
-            );
-        }
-    }
-
-    #[test]
-    fn unit_parse_bool_flag_rejects_false_like_values() {
-        for value in ["0", "false", "off", "no", ""] {
-            assert!(
-                !crate::drivers::shared_helpers::parse_bool_flag(value),
-                "expected falsey for {value}"
-            );
-        }
-    }
-
-    #[test]
-    fn unit_live_execution_enabled_from_env_honors_true_and_false_markers() {
-        with_env_vars(
-            &[
-                (SDK_DIRECT_LIVE_ENV, Some("1")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                assert!(
-                    live_execution_enabled_from_env(),
-                    "truthy env value should enable live SDK-direct mode",
-                );
-            },
-        );
-
-        with_env_vars(&[(SDK_DIRECT_LIVE_ENV, Some("0"))], || {
-            assert!(
-                !live_execution_enabled_from_env(),
-                "falsey env value should disable live SDK-direct mode",
-            );
-        });
-    }
-
-    #[test]
-    fn unit_run_live_s01_discovery_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("not-a-valid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-                ("KAMN_AGENT_NAME", Some("sdk-driver-test")),
-            ],
-            || {
-                let error =
-                    run_live_s01_discovery_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("service.endpoint") || error.contains("service endpoint"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s02_direct_message_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error =
-                    run_live_s02_direct_message_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s03_group_channel_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error =
-                    run_live_s03_group_channel_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s03_group_channel_probe_rejects_query_message_id_mismatch() {
-        let error = validate_live_s03_query_message_response("message-1", "message-2", "sent")
-            .expect_err("mismatched query message_id should fail");
-        assert!(
-            error.contains("mismatched message_id"),
-            "error should mention message_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s03_group_channel_probe_rejects_list_channel_id_mismatch() {
-        let error = validate_live_s03_list_messages_response("channel-1", "channel-2")
-            .expect_err("mismatched listed channel_id should fail");
-        assert!(
-            error.contains("mismatched channel_id"),
-            "error should mention channel_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s04_task_lifecycle_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("not-a-valid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-                ("KAMN_AGENT_NAME", Some("sdk-driver-test")),
-            ],
-            || {
-                let error =
-                    run_live_s04_task_lifecycle_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("service.endpoint") || error.contains("service endpoint"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s05_escrow_settlement_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s05_escrow_settlement_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_validate_live_s05_release_escrow_receipt_rejects_mismatched_escrow_id() {
-        let error = validate_live_s05_release_escrow_receipt("escrow-a", "escrow-b", "released")
-            .expect_err("mismatched escrow ids should fail");
-        assert!(
-            error.contains("mismatched escrow_id"),
-            "error should describe escrow-id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s06_proof_verification_probe_rejects_invalid_block_height_env_value() {
-        with_env_vars(
-            &[("KAMN_E2E_S06_PROOF_BLOCK_HEIGHT", Some("not-a-number"))],
-            || {
-                let error = run_live_s06_proof_verification_probe()
-                    .expect_err("invalid block height env value should fail");
-                assert!(
-                    error.contains("invalid block height env value"),
-                    "probe error should reflect parse failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s07_replay_protection_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s07_replay_protection_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s08_crash_recovery_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error =
-                    run_live_s08_crash_recovery_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s09_transport_failover_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                (
-                    "KAMN_E2E_S09_FAILOVER_ENDPOINT",
-                    Some("invalid-failover-endpoint"),
-                ),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s09_transport_failover_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s10_topology_coherence_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                (
-                    "KAMN_E2E_S10_PRIMARY_ENDPOINT",
-                    Some("invalid-primary-endpoint"),
-                ),
-                (
-                    "KAMN_E2E_S10_SECONDARY_ENDPOINT",
-                    Some("invalid-secondary-endpoint"),
-                ),
-                (
-                    "KAMN_E2E_S10_TERTIARY_ENDPOINT",
-                    Some("invalid-tertiary-endpoint"),
-                ),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s10_topology_coherence_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s11_signer_rotation_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error =
-                    run_live_s11_signer_rotation_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s12_retention_deletion_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s12_retention_deletion_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s13_bridge_forwarding_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s13_bridge_forwarding_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s14_batch_merkle_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error =
-                    run_live_s14_batch_merkle_probe().expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s15_performance_smoke_probe_rejects_invalid_endpoint() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("invalid-endpoint")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-            ],
-            || {
-                let error = run_live_s15_performance_smoke_probe()
-                    .expect_err("invalid endpoint should fail");
-                assert!(
-                    error.contains("connect failed"),
-                    "probe error should reflect connection failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_validate_s12_content_id_match_rejects_mismatch() {
-        let error = validate_s12_content_id_match("content-a", "content-b", "test step")
-            .expect_err("mismatched content ids should fail");
-        assert!(
-            error.contains("mismatched content_id"),
-            "error should mention content_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s12_content_field_coherence_rejects_drift() {
-        let error = validate_s12_content_field_coherence(
-            "tombstoned",
-            "expired",
-            "lifecycle_state",
-            "test step",
-        )
-        .expect_err("field drift should fail");
-        assert!(
-            error.contains("lifecycle_state drift"),
-            "error should mention field drift: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s13_bridge_id_match_rejects_mismatch() {
-        let error = validate_s13_bridge_id_match("bridge-a", "bridge-b", "test step")
-            .expect_err("mismatched bridge ids should fail");
-        assert!(
-            error.contains("mismatched bridge_id"),
-            "error should mention bridge_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s13_bridge_field_coherence_rejects_drift() {
-        let error =
-            validate_s13_bridge_field_coherence("forwarded", "stale", "bridge_status", "test step")
-                .expect_err("bridge field drift should fail");
-        assert!(
-            error.contains("bridge_status drift"),
-            "error should mention field drift: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s07_replay_reason_marker_accepts_expected_marker() {
-        validate_s07_replay_reason_marker(
-            "operation failed: service_api_auth_replay_nonce_detected",
-            "test helper",
-        )
-        .expect("expected marker should be accepted");
-    }
-
-    #[test]
-    fn unit_validate_s07_replay_reason_marker_rejects_missing_marker() {
-        let error = validate_s07_replay_reason_marker("operation failed", "test helper")
-            .expect_err("missing marker should fail");
-        assert!(
-            error.contains("missing replay reason marker"),
-            "error should mention replay marker contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s08_message_receipt_fields_rejects_empty_message_id() {
-        let error = validate_s08_message_receipt_fields("", "sent", "test helper")
-            .expect_err("empty message_id should fail");
-        assert!(
-            error.contains("empty message_id"),
-            "error should mention message_id requirement: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s08_query_message_response_rejects_mismatched_message_id() {
-        let error = validate_s08_query_message_response("message-1", "message-2", "sent", "test")
-            .expect_err("mismatched query message_id should fail");
-        assert!(
-            error.contains("mismatched message_id"),
-            "error should mention message_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s08_distinct_message_ids_accepts_distinct_ids() {
-        validate_s08_distinct_message_ids("message-1", "message-2", "test helper")
-            .expect("distinct message ids should pass");
-    }
-
-    #[test]
-    fn unit_validate_s08_distinct_message_ids_rejects_duplicate_ids() {
-        let error = validate_s08_distinct_message_ids("message-1", "message-1", "test helper")
-            .expect_err("duplicate message ids should fail");
-        assert!(
-            error.contains("duplicate message_id"),
-            "error should mention duplicate message_id: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_proof_response_accepts_valid_receipt() {
-        validate_s14_proof_response("message-1", "message-1", 42, "FINAL", true, "test helper")
-            .expect("valid S-14 proof response should pass");
-    }
-
-    #[test]
-    fn unit_validate_s14_proof_response_rejects_mismatched_message_id() {
-        let error =
-            validate_s14_proof_response("message-1", "message-2", 42, "FINAL", true, "test helper")
-                .expect_err("mismatched message_id should fail");
-        assert!(
-            error.contains("mismatched message_id"),
-            "error should mention message_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_proof_response_rejects_unverified_receipt() {
-        let error = validate_s14_proof_response(
-            "message-1",
-            "message-1",
-            42,
-            "FINAL",
-            false,
-            "test helper",
-        )
-        .expect_err("verified=false should fail");
-        assert!(
-            error.contains("verified=false"),
-            "error should mention verified contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_proof_response_rejects_non_final_finality() {
-        let error = validate_s14_proof_response(
-            "message-1",
-            "message-1",
-            42,
-            "PENDING",
-            true,
-            "test helper",
-        )
-        .expect_err("non-final finality should fail");
-        assert!(
-            error.contains("non-final finality"),
-            "error should mention finality contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_proof_response_rejects_zero_block_height() {
-        let error =
-            validate_s14_proof_response("message-1", "message-1", 0, "FINAL", true, "test helper")
-                .expect_err("block_height=0 should fail");
-        assert!(
-            error.contains("block_height=0"),
-            "error should mention block-height contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_accepts_within_budget_samples() {
-        validate_s15_latency_budget_samples(&[10, 20, 30], 80, 100, 25, 35, "test helper")
-            .expect("within-budget samples should pass");
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_rejects_total_budget_violation() {
-        let error =
-            validate_s15_latency_budget_samples(&[10, 20, 30], 120, 100, 25, 35, "test helper")
-                .expect_err("total budget violation should fail");
-        assert!(
-            error.contains("total elapsed millis exceeded budget"),
-            "error should mention total budget: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_rejects_p50_budget_violation() {
-        let error =
-            validate_s15_latency_budget_samples(&[10, 50, 90], 90, 200, 20, 100, "test helper")
-                .expect_err("p50 budget violation should fail");
-        assert!(
-            error.contains("p50 millis exceeded budget"),
-            "error should mention p50 budget: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_rejects_p99_budget_violation() {
-        let error =
-            validate_s15_latency_budget_samples(&[10, 20, 90], 90, 200, 50, 80, "test helper")
-                .expect_err("p99 budget violation should fail");
-        assert!(
-            error.contains("p99 millis exceeded budget"),
-            "error should mention p99 budget: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_parse_s15_budget_env_u128_uses_default_when_env_missing() {
-        with_env_vars(&[("KAMN_E2E_S15_TOTAL_BUDGET_MS", None)], || {
-            let parsed = parse_s15_budget_env_u128(
-                "KAMN_E2E_S15_TOTAL_BUDGET_MS",
-                91,
-                "sdk-direct live s15 test helper",
-            )
-            .expect("missing env key should use default");
-            assert_eq!(parsed, 91);
-        });
-    }
-
-    #[test]
-    fn unit_parse_s15_budget_env_u128_parses_positive_env_value() {
-        with_env_vars(&[("KAMN_E2E_S15_TOTAL_BUDGET_MS", Some("143"))], || {
-            let parsed = parse_s15_budget_env_u128(
-                "KAMN_E2E_S15_TOTAL_BUDGET_MS",
-                91,
-                "sdk-direct live s15 test helper",
-            )
-            .expect("valid env key should parse");
-            assert_eq!(parsed, 143);
-        });
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_accepts_exact_budget_boundaries() {
-        validate_s15_latency_budget_samples(&[10, 20, 30], 60, 60, 20, 30, "test helper")
-            .expect("equal total/p50/p99 budget boundaries should pass");
-    }
-
-    #[test]
-    fn unit_percentile_index_returns_expected_midpoint_index() {
-        assert_eq!(
-            crate::drivers::shared_helpers::percentile_index(3, 50),
-            1,
-            "len=3 and p50 should map to middle sample index",
-        );
-    }
-
-    #[test]
-    fn unit_percentile_index_clamps_percentile_above_hundred_to_last_index() {
-        assert_eq!(
-            crate::drivers::shared_helpers::percentile_index(3, 150),
-            2,
-            "percentiles above 100 should clamp to the last sample index",
-        );
-    }
-
-    #[test]
-    fn unit_live_s07_probe_agent_suffix_is_non_empty_numeric() {
-        let suffix = super::live_s07_probe_agent_suffix();
-        assert!(!suffix.is_empty(), "suffix should be non-empty");
-        assert!(
-            suffix.chars().all(|character| character.is_ascii_digit()),
-            "suffix should be numeric: {suffix}",
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s06_proof_verification_probe_accepts_final_verified_receipt() {
-        with_env_vars(
-            &[
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-                ("KAMN_KOLME_ENDPOINT", Some("http://localhost:3000")),
-                ("KAMN_AGENT_NAME", Some("sdk-driver-test")),
-                ("KAMN_E2E_S06_PROOF_FINALITY", Some("final")),
-            ],
-            || {
-                run_live_s06_proof_verification_probe()
-                    .expect("final verified proof probe should succeed");
-            },
-        );
-    }
-
-    #[test]
-    fn unit_sdk_direct_driver_debug_includes_live_toggle_field() {
-        let driver = SdkDirectDriver::with_probe(false, || Ok(()));
-        let debug = format!("{driver:?}");
-        assert!(
-            debug.contains("SdkDirectDriver"),
-            "debug output should include struct name: {debug}",
-        );
-        assert!(
-            debug.contains("live_execution_enabled"),
-            "debug output should include live toggle field: {debug}",
-        );
-    }
-
-    #[test]
-    fn spec_c00_live_disabled_driver_path_fails_closed_without_probe_invocation() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-        use std::sync::Arc;
-
-        let probe_calls = Arc::new(AtomicUsize::new(0));
-        let probe_calls_for_closure = Arc::clone(&probe_calls);
-        let driver = SdkDirectDriver::with_probe(false, move || {
-            probe_calls_for_closure.fetch_add(1, Ordering::SeqCst);
-            Ok(())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-01");
-        assert_eq!(
-            result.status, "fail",
-            "live-disabled S-01 must fail closed instead of reporting pass",
-        );
-        assert_eq!(
-            probe_calls.load(Ordering::SeqCst),
-            0,
-            "live probe should not be invoked when toggle is disabled",
-        );
-    }
-
-    #[test]
-    fn spec_c01_live_s04_driver_path_fails_closed_when_task_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s04 task probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-04");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-04 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c02_live_s06_driver_path_fails_closed_when_proof_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s06 proof probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-06");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-06 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c03_live_s02_driver_path_fails_closed_when_message_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s02 message probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-02");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-02 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c04_live_s03_driver_path_fails_closed_when_channel_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s03 channel probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-03");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-03 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c05_live_s05_driver_path_fails_closed_when_escrow_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s05 escrow probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-05");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-05 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c07_live_s07_driver_path_fails_closed_when_replay_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s07 replay probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-07");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-07 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c08_live_s08_driver_path_fails_closed_when_crash_recovery_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s08 crash-recovery probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-08");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-08 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c09_live_s09_driver_path_fails_closed_when_transport_failover_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s09 transport-failover probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-09");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-09 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c10_live_s10_driver_path_fails_closed_when_topology_coherence_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s10 topology-coherence probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-10");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-10 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c11_live_s11_driver_path_fails_closed_when_signer_rotation_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s11 signer-rotation probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-11");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-11 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c12_live_s12_driver_path_fails_closed_when_retention_deletion_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s12 retention-deletion probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-12");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-12 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c13_live_s13_driver_path_fails_closed_when_bridge_forwarding_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s13 bridge-forwarding probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-13");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-13 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c14_live_s14_driver_path_fails_closed_when_batch_merkle_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s14 batch-merkle probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-14");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-14 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c15_live_s15_driver_path_fails_closed_when_performance_smoke_probe_errors() {
-        let driver = SdkDirectDriver::with_probe(true, || {
-            Err("sdk-direct live s15 performance-smoke probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-15");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-15 should fail closed on probe error",
-        );
-    }
-}
+#[path = "sdk_direct_tests.rs"]
+mod sdk_direct_tests;
